🧠 Refactor : 사용하지 않는 컨트롤러 삭제
